requirements.txt
asn1crypto==0.24.0
Babel==2.6.0
cffi==1.11.5
cryptography==2.3.1
Django==1.11.3
django-environ==0.4.3
django-fernet-fields==0.5
django-phonenumber-field==1.3.0
idna==2.7
mysqlclient==1.3.13
phonenumberslite==8.9.11
pkg-resources==0.0.0
pycparser==2.18
pytz==2018.5
six==1.11.0
